requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Task queue
celery==5.3.4
redis==5.0.1

# Configuration
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Security
python-magic==0.4.27
cryptography==41.0.8

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Monitoring
prometheus-client==0.19.0
psutil==5.9.6

# Utilities
PyYAML==6.0.1
starlette==0.27.0
jinja2==3.1.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8